requirements.txt
django-extensions==3.2.3
Django
asgiref
sqlparse
tzdata
djoser==2.3.1
djangorestframework_simplejwt
PyJWT
social-auth-app-django
social-auth-core
cryptography
cffi
pycparser
defusedxml
oauthlib
python3-openid
requests
certifi
charset-normalizer
idna
urllib3
requests-oauthlib
drf-spectacular==0.28.0
inflection
jsonschema
attrs
jsonschema-specifications
referencing
rpds-py
typing_extensions
PyYAML
uritemplate
drf-spectacular-sidecar==2025.2.1
gunicorn==23.0.0
packaging
psycopg2==2.9.10
whitenoise==6.8.2
